Fix + Remove (#88)
I fixed a small issue with how we describe what one of the functions returns. It now accurately reflects what it does.
I removed an unnecessary bit of code that wasn't doing anything useful.
I improved the readability of one function by changing how a long line is split up. It now looks cleaner and is easier to read